+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * (theimo1221) Fix unit Test
 * (theimo1221) Update packages
 * (theimo1221) Add Humidity Sensor Persisting
+* (theimo1221) Clear Persisting from old mixed persisting commands
 
 ## 1.1.48 (2022-10-15)